requirements.txt
alabaster==0.7.7
argon2-cffi==19.2.0
asgiref==3.3.1
autopep8==1.4.4
Babel==2.2.0
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
cssselect==0.9.1
defusedxml==0.6.0
Django==2.2.14
django-allauth==0.39.1
django-braces==1.14.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-js-asset==1.2.2
django-livesync==0.5
django-multiselectfield==0.1.12
django-utils-six==2.0
djangorestframework==3.11.0
Faker==4.0.3
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
Markdown==2.6.5
oauthlib==3.0.1
pathtools==0.1.2
pep8==1.7.1
Pillow==6.2.2
pycodestyle==2.5.0
pycparser==2.20
python-dateutil==1.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.14.0
sqlparse==0.2.4
stripe==2.27.0
text-unidecode==1.3
tornado==6.0.4
urllib3==1.24.2
watchdog==0.10.3
websocket-client==0.57.0
wincertstore==0.2
